Added check to ensure that deletion does not remove last existing contact detail

+++ spring-backend/src/main/java/org/example/contact/ContactController.java
@@ -58,7 +58,10 @@ public class ContactController {
 
     @DeleteMapping(path="/contact/{id}")
     public ResponseEntity<Void> deleteContact(@NotBlank @Digits(integer = 19, fraction = 0) @PathVariable("id") final String id) {
-        //TODO: validate that it is NOT the last contact associated with account else reject removal
+        if (!contactService.canRemoveContactDetails(id)) {
+            return new ResponseEntity<>(HttpStatusCode.valueOf(400));
+        }
+
         contactService.deleteById(id);
         return new ResponseEntity<>(HttpStatusCode.valueOf(204));
     }
